refactor: move isEssentialsFlow to client-safe utils/common module
Agent-Logs-Url: https://github.com/edx/frontend-app-enterprise-checkout/sessions/524d8165-8753-4d83-b292-a9aeb878c937
