+++ src/components/billing-details-pages/BillingDetailsPage.tsx
@@ -24,7 +24,7 @@ import EVENT_NAMES from '@/constants/events';
 import { useCheckoutFormStore, useCurrentPageDetails } from '@/hooks/index';
 import { sendEnterpriseCheckoutTrackingEvent } from '@/utils/common';
 
-import { isEssentialsFlow } from '../app/routes/loaders/utils';
+import { isEssentialsFlow } from '@/utils/common';
 
 const BillingDetailsPage: React.FC = () => {
   const navigate = useNavigate();
